+++ src/main/java/com/umc/EveryWear/domain/review/service/query/ReviewQueryServiceImpl.java
@@ -1,7 +1,6 @@
 package com.umc.EveryWear.domain.review.service.query;
 
 import com.umc.EveryWear.domain.product.entity.Product;
-import com.umc.EveryWear.domain.product.entity.ReviewCrawlStatus;
 import com.umc.EveryWear.domain.product.repository.ProductRepository;
 import com.umc.EveryWear.domain.review.dto.res.ReviewResDTO;
 import com.umc.EveryWear.domain.review.entity.Review;
